Created initial plugin to gather data from the game and use that data to control the game
Created an angelscript file that creates a menu item used to start the update function. Within the update function it creates all of the variables that are used to gather data from the game and send it to a socket. I then on a python script create two threads one will listen to the socket and send the data to a file as well as update a global variable that the other thread can use. The other thread reads parts of the global variable and enacts certain effects in the game. Currently all it does is keep itself at 50 m/s.